@@ -1,39 +1,34 @@
 {
-  "name": "blog",
-  "version": "1.0.0",
+  "name": "client",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.19.0",
-    "d3": "^5.12.0",
-    "highlight.js": "^9.16.2",
-    "just-detect-adblock": "^1.0.0",
-    "moment": "^2.24.0",
-    "react": "^16.11.0",
-    "react-dom": "^16.11.0",
-    "react-fontawesome": "^1.6.1",
-    "react-helmet": "^5.2.1",
-    "react-markdown": "^4.2.2",
-    "react-redux": "^7.1.1",
-    "react-router": "^5.1.2",
-    "react-scripts": "3.2.0",
-    "react-toastify": "^6.0.8",
-    "react-virtualized-auto-sizer": "^1.0.2",
-    "redux": "^4.0.4",
-    "redux-persist": "^6.0.0",
-    "retry": "^0.12.0"
+    "axios": "^0.27.2",
+    "d3": "^7.5.0",
+    "moment": "^2.29.3",
+    "node-sass": "^7.0.1",
+    "prismjs": "^1.28.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-helmet": "^6.1.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.1",
+    "rxjs": "^7.5.5",
+    "showdown": "^2.1.0",
+    "trie-search": "^1.3.6"
   },
   "scripts": {
-    "build-css": "node-sass-chokidar src/ -o src/",
-    "watch-css": "npm run build-css && node-sass-chokidar src/ -o src/ --watch --recursive",
-    "start-js": "react-scripts start",
-    "start": "npm-run-all -p watch-css start-js",
-    "build-js": "react-scripts build",
-    "build": "npm-run-all build-css build-js",
-    "test": "react-scripts test --env=jsdom",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -46,10 +41,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "node-sass-chokidar": "^1.3.5",
-    "npm-run-all": "^4.1.5",
-    "redux-logger": "^3.0.6"
   }
 }
