@@ -4,31 +4,36 @@
   "description": "",
   "main": "src/index.js",
   "engines": {
-    "node": "10.19.0"
+    "node": ">=16.10"
   },
+  "engineStrict": true,
   "scripts": {
-    "start": "nodemon src/index.js",
-    "util": "node src/util.js"
+    "start": "nodemon src/index.js"
   },
-  "author": "Jack Mead",
+  "author": "",
   "license": "ISC",
+  "_moduleAliases": {
+    "@middleware": "./src/middleware",
+    "@clients": "./src/clients",
+    "@config": "./src/config.js"
+  },
   "dependencies": {
-    "axios": "^0.19.0",
-    "body-parser": "^1.19.0",
+    "ajv": "^8.11.0",
+    "ajv-formats": "^2.1.1",
+    "async-retry": "^1.3.3",
+    "chokidar": "^3.5.3",
     "compression": "^1.7.4",
     "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "finviz": "0.0.3",
-    "googleapis": "^47.0.0",
-    "js-base64": "^2.5.2",
-    "node-cache": "^5.0.1",
-    "nodemailer": "^6.4.2",
-    "nodemon": "^1.19.4",
-    "pg": "^7.12.1",
-    "retry": "^0.12.0",
-    "showdown": "^1.9.1"
-  },
-  "devDependencies": {
-    "xml": "^1.0.1"
+    "express": "^4.18.1",
+    "fuse.js": "^6.6.2",
+    "helmet": "^5.1.0",
+    "module-alias": "^2.2.2",
+    "moment": "^2.29.3",
+    "node-cache": "^5.1.2",
+    "nodemon": "^2.0.16",
+    "pg": "^8.7.3",
+    "trie-search": "^1.3.6",
+    "winston": "^3.7.2",
+    "yaml": "^2.1.1"
   }
 }
