@@ -1,10 +1,8 @@
 import React, { Component } from 'react'
-import * as three from 'three';
 import * as d3 from 'd3';
 import { dragElement } from './util';
 
 import JSEngine from './jsengine';
-import { timeHours } from 'd3';
 let RSEngine = null;
 
 function randomColor() {
@@ -17,7 +15,6 @@ function randomColor() {
 }
 
 const d3Rects = {};
-const tjRects = {};
 
 export default class App extends Component {
   constructor(props) {
@@ -35,7 +32,6 @@ export default class App extends Component {
       gravityConstant: 0.0005,
       damping: 1,
       fps: 0,
-      d3: true,
     };
 
     this.animation = null;
@@ -47,7 +43,6 @@ export default class App extends Component {
     this.restart = this.restart.bind(this);
     this.center = this.center.bind(this);
     this.engine = null;
-    this.tj = {};
   }
 
   componentDidMount() {
@@ -75,8 +70,6 @@ export default class App extends Component {
   restart() {
     this.stop();
     this.d3Rects = {};
-    this.tjRects = {};
-    this.tj = {};
     const graph = document.getElementById('graph');
     if (graph) {
       for(const child of graph.children) {
@@ -91,11 +84,7 @@ export default class App extends Component {
 
   start() {
     if (this.animation === null) {
-      if (this.state.d3) {
-        this.d3Loop();
-      } else {
-        this.tjLoop();
-      }
+      this.d3Loop();
     } 
   }
 
@@ -114,11 +103,7 @@ export default class App extends Component {
   }
 
   createEngine() {
-    if (this.state.d3) {
-      this.createD3Engine();
-    } else {
-      return this.createTJEngine();
-    }
+    this.createD3Engine();
   }
 
   createD3Engine() {
@@ -156,38 +141,6 @@ export default class App extends Component {
         .attr('height', rect.height)
         .attr('x', rect.x)
         .attr('y', rect.y);
-    }
-  }
-
-  createTJEngine() {
-    const graph = document.getElementById('graph');
-    const bb = graph.getBoundingClientRect();
-    this.world_width = bb.width;
-    this.world_height = bb.height;
-    this.tj.camera = new three.PerspectiveCamera(75,this.world_width/this.world_height,0.01,10);
-    this.tj.camera.position.z = 1;
-    this.tj.scene = new three.Scene();
-    this.tj.renderer = new three.WebGLRenderer({ antialias: true });
-    this.tj.renderer.setSize( this.world_width, this.world_height );
-    graph.appendChild(this.tj.renderer.domElement);
-
-    const EngineFactory = this.state.wasm ? RSEngine : JSEngine;
-    this.engine = EngineFactory.new(
-      this.world_width,
-      this.world_height,
-      this.state.gravityConstant,
-      this.state.damping
-    );
-    this.engine.generate(this.state.amount);
-
-    const rects = this.engine.get_rects();
-    for (const rect of rects) {
-      const geometry = new three.BoxGeometry( 0.2, 0.2, 0.2 );
-      const material = new three.MeshNormalMaterial({color: 0x00ff00});
-      const mesh = new three.Mesh( geometry, material );
-      tjRects[rect.id] = mesh;
-      this.tj.scene.add(mesh);
-      break;
     }
   }
 
@@ -277,33 +230,6 @@ export default class App extends Component {
     this.animation = window.requestAnimationFrame(update);
   }
 
-  tjLoop() {
-    let fps = 0;
-    let last_fps = performance.now();
-    let update_arr = new Array(this.state.amount);
-
-    const update = () => {
-      this.animation = window.requestAnimationFrame(update);
-
-      this.engine.tick(update_arr);
-
-      for (const rect of update_arr) {
-        //tjRects[rect.id].rotation.x = 0.01;
-        //tjRects[rect.id].rotation.y = 0.01;
-      }
-
-      fps+=1;
-      if (performance.now() - last_fps > 1000) {
-        this.setState({ fps });
-        fps = 0;
-        last_fps = performance.now();
-      }
-
-      this.tj.renderer.render(this.tj.scene, this.tj.camera);
-    }
-    this.animation = window.requestAnimationFrame(update);
-  }
-
   renderControls() {
     return (
       <div className="controls">
@@ -315,9 +241,6 @@ export default class App extends Component {
         </button>
         <button onClick={this.restart}>
           Restart
-        </button>
-        <button onClick={this.center}>
-          Center
         </button>
         <div className="check">
           <p>Amount</p>
@@ -356,14 +279,6 @@ export default class App extends Component {
             style={{width: 100}}
             value={this.state.damping}
             onChange={this.onChangeDamping.bind(this)}
-          />
-        </div>
-        <div className="check">
-          {this.state.d3 ? 'D3' : 'TJ'}
-          <input
-            type="checkbox"
-            checked={this.state.d3}
-            onChange={() => this.setState({d3: !this.state.d3})}
           />
         </div>
         <div className="check">
